refec:move mongo config class for unit test and rename
